app/requirements.txt
Django>=3.0.7,<4
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework
django-cors-headers==3.4.0
drf-yasg